+++ sensor/contiki/sensor/subscriber.h
@@ -30,8 +30,8 @@ typedef union _cond_value_u {
 } cond_value_u;
 
 typedef struct _condition_t {
-    uint32_t period;
-    uint32_t expire_time;
+    uint16_t period;
+    uint16_t expire_time;
     enum operation_e operation;
     cond_value_u value;
 } condition_t;
@@ -53,10 +53,10 @@ res_subscriber_t *subscriber_alloc();
 void subscriber_free(res_subscriber_t *subscriber);
 
 int32_t subscriber_period_type_init(res_subscriber_t *subscriber, uip_ip6addr_t *addr, 
-                                    uint8_t *device_id, uint32_t period);
+                                    uint8_t *device_id, uint16_t period);
 
 int32_t subscriber_expire_type_init(res_subscriber_t *subscriber, uip_ip6addr_t *addr,
-                                    uint8_t *device_id, uint32_t expire_time);
+                                    uint8_t *device_id, uint16_t expire_time);
 
 int32_t subscriber_value_type_init(res_subscriber_t *subscriber, uip_ip6addr_t *addr, 
                                    uint8_t *device_id, enum operation_e operation, cond_value_u *value);
